% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareek_scale.R
\name{pareek_scale}
\alias{pareek_scale}
\title{Udai Pareek Socioeconomic Status Scale}
\usage{
pareek_scale(caste, occupation, education, land, social.participation, house,
  farm.power, material.possessions, family.type)
}
\arguments{
\item{caste}{A character, Caste of the family, Possible values are - "schedule caste","lower caste","artisan caste","agriculture caste","prestige caste","dominant caste"}

\item{occupation}{A character, Occupation of head of family, Possible values are - "none","laborer","caste occupation","business","independent profession","cultivation","service"}

\item{education}{A character,Education of head of family, Possible values are - "illiterate","can read only","can read and write","primary","middle","high school","graduate","above graduate"}

\item{land}{A number, Land owned in acres}

\item{social.participation}{A character, Social participation of family, Possible values are - "none","member of one organization","member of more than one organization",
"office holder in such a organization","wide public leader"}

\item{house}{A character, House type, Possible values are - "no house","hut","kutcha house","mixed house","pucca house","mansion"}

\item{farm.power}{A number, Number of draught animals}

\item{material.possessions}{A character, Material possesions, Possible values are - "bullock cart","cycle","radio","chair","mobile phone","television","refrigerator"}

\item{family.type}{A character, Family type, Possible values are "single","joint","extended","size up to 5","any other distinctive features"}
}
\value{
Udai Pareek Socioeconomic Status Class.
}
\description{
This function measures the socioeconomic status according to Udai Pareek Scale.
}
\details{
This function measures the socioeconomic status according to Udai Pareek Scale based on three parameters given to it (occupation,education,income).
}
\author{
Biswarup Pramanik
}
