% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prasad_scale.R
\name{prasad_scale}
\alias{prasad_scale}
\title{Realtime BG Prasad Socioeconomic Status Scale}
\usage{
prasad_scale(income)
}
\arguments{
\item{income}{A number, Family Income per month in Rupees}
}
\value{
BG Prasad Socioeconomic Class.
}
\description{
This function measures the realtime socioeconomic status according to BG Prasad Scale.
}
\details{
This function measures the realtime socioeconomic status according to BG Prasad Scale based on income parameter given to it.
NOTE: You will be asked to enter the latest Consumer Price Index-Industrial Worker (CPI-IW) value
available from the website "http://labourbureaunew.gov.in/LBO_indnum.htm"
}
\author{
Biswarup Pramanik
}
